// cluster/console/console/web/package/src/apis/cordiumv1/cordiumv1.ts
export interface Space {
    /**
     * APIVersion is the API version (i.e. "cordium/v1")
     *
     * @generated from protobuf field: string apiVersion = 1
     */
    apiVersion: string;
    /**
     * Kind is the resource name (i.e. `Space`).
     *
     * @generated from protobuf field: string kind = 2
     */
    kind: string;
    /**
     * Metadata is the object's metadata.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.Metadata metadata = 3
     */
    metadata?: Metadata;
    /**
     * Spec is the Space specification.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Space.Spec spec = 4
     */
    spec?: Space_Spec;
    /**
     * Status is the current status of the Space.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Space.Status status = 5
     */
    status?: Space_Status;
}
/**
 * Spec is the Space specification
 *
 * @generated from protobuf message octelium.api.main.cordium.v1.Space.Spec
 */
export interface Space_Spec {
    /**
     * Limit is the default and the maximum compute resources of the Space's
     * Workspaces.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Space.Spec.Limit limit = 1
     */
    limit?: Space_Spec_Limit;
    /**
     * Runtime is the runtime configuration that cascades down to every
     * Workspace of the Space.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Space.Spec.Runtime runtime = 2
     */
    runtime?: Space_Spec_Runtime;
    /**
     * Authorization is the access control configuration of the Space's
     * Workspaces.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Space.Spec.Authorization authorization = 3
     */
    authorization?: Space_Spec_Authorization;
}
/**
 * Limit is the default and the maximum compute resources of the Space's
 * Workspaces. It can only be set for ORGANIZATION Spaces.
 *
 * @generated from protobuf message octelium.api.main.cordium.v1.Space.Spec.Limit
 */
export interface Space_Spec_Limit {
    /**
     * DefaultLimit is the limits that are applied to the Workspaces of the
     * Space that do not define their own limits.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Limit defaultLimit = 1
     */
    defaultLimit?: Workspace_Spec_Limit;
    /**
     * MaxLimit is a hard cap that no Workspace of the Space can exceed.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Limit maxLimit = 2
     */
    maxLimit?: Workspace_Spec_Limit;
}
/**
 * Runtime is the runtime configuration that cascades down to every
 * Workspace of the Space regardless of its Template.
 *
 * @generated from protobuf message octelium.api.main.cordium.v1.Space.Spec.Runtime
 */
export interface Space_Spec_Runtime {
    /**
     * EnvVars is the list of the environment variables that are injected
     * into every Workspace of the Space. Secret-sourced environment
     * variables can only be used in ORGANIZATION Spaces.
     *
     * @generated from protobuf field: repeated octelium.api.main.cordium.v1.Workspace.Spec.Runtime.EnvVar envVars = 1
     */
    envVars: Workspace_Spec_Runtime_EnvVar[];
    /**
     * Tasks is the list of the lifecycle tasks that are run in every
     * Workspace of the Space.
     *
     * @generated from protobuf field: repeated octelium.api.main.cordium.v1.Workspace.Spec.Runtime.Task tasks = 2
     */
    tasks: Workspace_Spec_Runtime_Task[];
    /**
     * Capabilities is the Linux capabilities that are merged into every
     * Workspace of the Space.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Runtime.Capabilities capabilities = 3
     */
    capabilities?: Workspace_Spec_Runtime_Capabilities;
}
/**
 * Authorization is the access control configuration of the Space's
 * Workspaces.
 *
 * @generated from protobuf message octelium.api.main.cordium.v1.Space.Spec.Authorization
 */
export interface Space_Spec_Authorization {
    /**
     * DisableSSH denies SSH access to the Workspaces of the Space.
     *
     * @generated from protobuf field: bool disableSSH = 1
     */
    disableSSH: boolean;
}
/**
 * Status is the current status of the Space. It is managed by the Cluster
 * and it is read-only.
 *
 * @generated from protobuf message octelium.api.main.cordium.v1.Space.Status
 */
export interface Space_Status {
    /**
     * UserRef is the reference of the Octelium User who created the Space.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ObjectReference userRef = 1
     */
    userRef?: ObjectReference;
    /**
     * Type is the type of the Space.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Space.Status.Type type = 2
     */
    type: Space_Status_Type;
}
/**
 * Type is the type of the Space
 *
 * @generated from protobuf enum octelium.api.main.cordium.v1.Space.Status.Type
 */
export enum Space_Status_Type {
    /**
     * SPACE_TYPE_UNSET is not used.
     *
     * @generated from protobuf enum value: SPACE_TYPE_UNSET = 0;
     */
    SPACE_TYPE_UNSET = 0,
    /**
     * USER is a Space that is personal to a single Octelium User. Members
     * cannot currently be added to it and it cannot define resource limits.
     *
     * @generated from protobuf enum value: USER = 1;
     */
    USER = 1,
    /**
     * ORGANIZATION is a shared Space that can have several Members with
     * different roles.
     *
     * @generated from protobuf enum value: ORGANIZATION = 2;
     */
    ORGANIZATION = 2
}
// ...
class Space$Type extends MessageType<Space> {
    constructor() {
        super("octelium.api.main.cordium.v1.Space", [
            { no: 1, name: "apiVersion", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "kind", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "metadata", kind: "message", T: () => Metadata },
            { no: 4, name: "spec", kind: "message", T: () => Space_Spec },
            { no: 5, name: "status", kind: "message", T: () => Space_Status }
        ]);
    }
    create(value?: PartialMessage<Space>): Space {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.apiVersion = "";
        message.kind = "";
        if (value !== undefined)
            reflectionMergePartial<Space>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Space): Space {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string apiVersion */ 1:
                    message.apiVersion = reader.string();
                    break;
                case /* string kind */ 2:
                    message.kind = reader.string();
                    break;
                case /* octelium.api.main.meta.v1.Metadata metadata */ 3:
                    message.metadata = Metadata.internalBinaryRead(reader, reader.uint32(), options, message.metadata);
                    break;
                case /* octelium.api.main.cordium.v1.Space.Spec spec */ 4:
                    message.spec = Space_Spec.internalBinaryRead(reader, reader.uint32(), options, message.spec);
                    break;
                case /* octelium.api.main.cordium.v1.Space.Status status */ 5:
                    message.status = Space_Status.internalBinaryRead(reader, reader.uint32(), options, message.status);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Space, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string apiVersion = 1; */
        if (message.apiVersion !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.apiVersion);
        /* string kind = 2; */
        if (message.kind !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.kind);
        /* octelium.api.main.meta.v1.Metadata metadata = 3; */
        if (message.metadata)
            Metadata.internalBinaryWrite(message.metadata, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* octelium.api.main.cordium.v1.Space.Spec spec = 4; */
        if (message.spec)
            Space_Spec.internalBinaryWrite(message.spec, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        /* octelium.api.main.cordium.v1.Space.Status status = 5; */
        if (message.status)
            Space_Status.internalBinaryWrite(message.status, writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.cordium.v1.Space
 */
export const Space = new Space$Type();
// ...
export const Space_Spec = new Space_Spec$Type();
// ...
export const Space_Spec_Limit = new Space_Spec_Limit$Type();
// ...
export const Space_Spec_Runtime = new Space_Spec_Runtime$Type();
// ...
export const Space_Spec_Authorization = new Space_Spec_Authorization$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Space_Status$Type extends MessageType<Space_Status> {
    constructor() {
        super("octelium.api.main.cordium.v1.Space.Status", [
            { no: 1, name: "userRef", kind: "message", T: () => ObjectReference },
            { no: 2, name: "type", kind: "enum", T: () => ["octelium.api.main.cordium.v1.Space.Status.Type", Space_Status_Type] }
        ]);
    }
    create(value?: PartialMessage<Space_Status>): Space_Status {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.type = 0;
        if (value !== undefined)
            reflectionMergePartial<Space_Status>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Space_Status): Space_Status {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* octelium.api.main.meta.v1.ObjectReference userRef */ 1:
                    message.userRef = ObjectReference.internalBinaryRead(reader, reader.uint32(), options, message.userRef);
                    break;
                case /* octelium.api.main.cordium.v1.Space.Status.Type type */ 2:
                    message.type = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Space_Status, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* octelium.api.main.meta.v1.ObjectReference userRef = 1; */
        if (message.userRef)
            ObjectReference.internalBinaryWrite(message.userRef, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* octelium.api.main.cordium.v1.Space.Status.Type type = 2; */
        if (message.type !== 0)
            writer.tag(2, WireType.Varint).int32(message.type);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.cordium.v1.Space.Status
 */
export const Space_Status = new Space_Status$Type();